added node::object
All relevant libnodecc objects are now managed using node::object which
provides an intrusive reference counter. This will help in preventing
the deletion of objects during asynchronous operations in libuv.

+++ include/libnodecc/buffer/buffer.h
@@ -210,7 +210,7 @@ protected:
 		void release();
 
 		const void* base;
-		std::atomic<uintptr_t> use_count;
+		std::atomic<unsigned int> use_count;
 	};
 
 	template<typename D>
